services/nutri_ai_service/requirements.txt
Flask
Werkzeug
requests
python-dotenv
easyocr
pillow
numpy
torch
torchvision
Jinja2
itsdangerous
gunicorn